classic meteor added

+++ SugarComa/Assets/MiniGames/FallingStars/Scripts/Player/PlayerSpecs.cs
@@ -6,7 +6,7 @@ namespace Assets.MiniGames.FallingStars.Scripts.Player
 {
     public class PlayerSpecs : MonoBehaviour
     {
-        public int _health;
+        public float _health;
         public bool _isDead;
         public float _moveSpeed = 5f;
         public float _rotationSpeed = 10f;
